Adds an option to show/hide tabs in the playground, and refactors the DS to be it's own thing
