+++ packages/typescriptlang-org/src/copy/en/playground.ts
@@ -6,6 +6,7 @@ export const playCopy = {
   play_subnav_examples_close: "Close",
   play_subnav_whatsnew: "What's New",
   play_subnav_settings: "Settings",
+  play_settings_tabs_settings: "Sidebar Tabs",
   play_downloading_typescript: "Downloading TypeScript...", // when loading
   play_downloading_version: "Version...", // when loading
   play_toolbar_run: "Run",
@@ -61,6 +62,23 @@ console.log(anExampleVariable)
 // To learn more about the language, click above in "Examples" or "What's New".
 // Otherwise, get started by removing these comments and the world is your playground.
   `,
+
+  play_sidebar_js_title: "JavaScript",
+  play_sidebar_js_blurb: "Shows the transpiled JS",
+
+  play_sidebar_dts_title: "Definition Files",
+  play_sidebar_dts_blurb: "Shows the .d.ts output of your code",
+
+  play_sidebar_err_title: "Compiler Errors",
+  play_sidebar_err_blurb: "Shows compiler errors in full",
+
+  play_sidebar_run_title: "Run JavaScript in Browser",
+  play_sidebar_run_blurb:
+    "Shows the output of running the JavaScript in the editor",
+
+  play_sidebar_plugins_title: "Manage Playground Plugins",
+  play_sidebar_plugins_blurb:
+    "Handles adding/removing 3rd party extensions to the playground",
 
   // Notes:
   // Compiler flag information is all from the tsconfig reference info
